requirements.txt
docker==7.1.0
fastapi==0.115.6
fastapi-cli==0.0.7
fastjsonschema==2.19.1
GitPython==3.1.43
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
python-json-logger==2.0.7
python-multipart==0.0.20
redis==5.2.1
uvicorn==0.34.0
virtualenv==20.28.1
psycopg2-binary==2.9.10
SQLAlchemy==2.0.39
sqlalchemy-orm==1.2.10
passlib==1.7.4
